requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
numpy